requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0